Move materialization tests to separate file
